requirements.txt
requests
pyyaml
python-dotenv
typer
rich
ruamel.yaml
anthropic
pydantic
fastapi
uvicorn
jinja2
google-genai
matplotlib
plotly
